@@ -7,56 +7,56 @@ publish_to: "none"
 version: 0.0.5+1
 
 environment:
-  sdk: ">=2.18.5 <3.0.0"
+  sdk: ">=2.19.2 <3.0.0"
 
 dependencies:
   another_flushbar: 1.12.29
-  badges: 2.0.3
+  badges: 3.0.2
   bot_toast: 4.0.3
-  collection: 1.16.0
-  connectivity_plus: 3.0.2
+  collection: 1.17.0
+  connectivity_plus: 3.0.3
   dartx: 1.1.0
   dartz: 0.10.1
-  data_table_2: 2.3.8
+  data_table_2: 2.3.12
   dio: 4.0.6
-  drift: 2.3.0
+  drift: 2.5.0
   easy_localization: 3.0.1
   ffi: 2.0.1
-  file_picker: 5.2.3
-  fl_chart: 0.55.2
-  fluent_ui: 4.0.3+1
+  file_picker: 5.2.5
+  fl_chart: 0.60.0
+  fluent_ui: 4.3.0
   flutter:
     sdk: flutter
-  flutter_riverpod: 2.1.1
-  flutter_screenutil: 5.6.0
-  font_awesome_flutter: 10.3.0
+  flutter_riverpod: 2.1.3
+  flutter_screenutil: 5.6.1
+  font_awesome_flutter: 10.4.0
   freezed_annotation: 2.2.0
-  go_router: 5.2.1
+  go_router: 6.0.6
   intl: 0.17.0
-  json_annotation: 4.7.0
+  json_annotation: 4.8.0
   logger: 1.1.0
-  modal_bottom_sheet: 2.1.2
-  package_info_plus: 3.0.2
+  modal_bottom_sheet: 3.0.0-pre
+  package_info_plus: 3.0.3
   path: 1.8.2
-  path_provider: 2.0.11
+  path_provider: 2.0.12
   pretty_dio_logger: 1.2.0-beta-1
   retrofit: 3.3.1
-  riverpod: 2.1.1
-  shared_preferences: 2.0.15
+  riverpod: 2.1.3
+  shared_preferences: 2.0.17
   sqlite3_flutter_libs: 0.5.12
   time: 2.1.3
-  url_launcher: 6.1.7
+  url_launcher: 6.1.9
   webview_windows: 0.2.1
-  win32: 3.1.2
-  window_manager: 0.2.8
+  win32: 3.1.3
+  window_manager: 0.3.0
 
 dev_dependencies:
-  build_runner: 2.3.2
-  drift_dev: 2.3.2
+  build_runner: 2.3.3
+  drift_dev: 2.5.2
   flutter_test:
     sdk: flutter
-  freezed: 2.3.0
-  json_serializable: 6.5.4
+  freezed: 2.3.2
+  json_serializable: 6.6.1
   lint: 2.0.1
   retrofit_generator: 4.2.0
 
