@@ -27,7 +27,7 @@ int APIENTRY wWinMain(_In_ HINSTANCE instance, _In_opt_ HINSTANCE prev,
   FlutterWindow window(project);
   Win32Window::Point origin(10, 10);
   Win32Window::Size size(1920, 1080);
-  if (!window.CreateAndShow(L"Arknights Analysis", origin, size)) {
+  if (!window.Create(L"Arknights Analysis", origin, size)) {
     return EXIT_FAILURE;
   }
   window.SetQuitOnClose(true);
